+++ rs/nervous_system/common/validation/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@rules_rust//rust:defs.bzl", "rust_library")
+
+package(default_visibility = ["//visibility:public"])
+
+rust_library(
+    name = "validation",
+    srcs = glob(["src/**"]),
+    crate_name = "ic_nervous_system_common_validation",
+    version = "0.9.0",
+    deps = [
+        # Keep sorted.
+    ],
+)
