+++ rs/nns/governance/api/BUILD.bazel
@@ -7,6 +7,7 @@ BASE_DEPENDENCIES = [
     # Keep sorted.
     "//rs/crypto/sha2",
     "//rs/nervous_system/clients",
+    "//rs/nervous_system/common/validation",
     "//rs/nervous_system/proto",
     "//rs/nns/common",
     "//rs/protobuf",
