+++ rs/nns/governance/api/Cargo.toml
@@ -17,6 +17,7 @@ comparable = { version = "0.5", features = ["derive"] }
 ic-base-types = { path = "../../../types/base_types" }
 ic-crypto-sha2 = { path = "../../../crypto/sha2" }
 ic-nervous-system-clients = { path = "../../../nervous_system/clients" }
+ic-nervous-system-common-validation = { path = "../../../nervous_system/common/validation" }
 ic-nervous-system-proto = { path = "../../../nervous_system/proto" }
 ic-nns-common = { path = "../../common" }
 ic-protobuf = { path = "../../../protobuf" }
