+++ rs/nns/governance/api/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bitcoin;
 pub mod pb;
-pub mod proposal_helpers;
+pub mod proposal_submission_helpers;
+pub mod proposal_validation;
 pub mod subnet_rental;
 pub mod test_api;
