+++ rs/nns/governance/src/governance/tests/mod.rs
@@ -13,6 +13,7 @@ use ic_nervous_system_common::{assert_is_err, assert_is_ok, E8};
 #[cfg(feature = "test")]
 use ic_nervous_system_proto::pb::v1::GlobalTimeOfDay;
 use ic_nns_common::pb::v1::NeuronId;
+use ic_nns_governance_api::pb::v1::CreateServiceNervousSystem as ApiCreateServiceNervousSystem;
 use ic_protobuf::registry::dc::v1::DataCenterRecord;
 #[cfg(feature = "test")]
 use ic_sns_init::pb::v1::SnsInitPayload;
@@ -248,9 +249,10 @@ mod convert_from_create_service_nervous_system_to_sns_init_payload_tests {
         // Step 1: Prepare the world. (In this case, trivial.)
 
         // Step 2: Call the code under test.
-        let converted =
-            SnsInitPayload::try_from(CREATE_SERVICE_NERVOUS_SYSTEM_WITH_MATCHED_FUNDING.clone())
-                .unwrap();
+        let converted = SnsInitPayload::try_from(ApiCreateServiceNervousSystem::from(
+            CREATE_SERVICE_NERVOUS_SYSTEM_WITH_MATCHED_FUNDING.clone(),
+        ))
+        .unwrap();
 
         // Step 3: Inspect the result.
 
@@ -498,7 +500,7 @@ mod convert_from_create_service_nervous_system_to_sns_init_payload_tests {
             });
 
         // Step 2: Call the code under test.
-        let converted = SnsInitPayload::try_from(original);
+        let converted = SnsInitPayload::try_from(ApiCreateServiceNervousSystem::from(original));
 
         // Step 3: Inspect the result: Err must contain "wait for quiet".
         match converted {
@@ -576,7 +578,10 @@ mod convert_create_service_nervous_system_proposal_to_sns_init_payload_tests_wit
                 .expect("Cannot compute swap_start_timestamp_seconds, swap_due_timestamp_seconds.")
             };
 
-            let sns_init_payload = SnsInitPayload::try_from(create_service_nervous_system).unwrap();
+            let sns_init_payload = SnsInitPayload::try_from(ApiCreateServiceNervousSystem::from(
+                create_service_nervous_system,
+            ))
+            .unwrap();
 
             SnsInitPayload {
                 neurons_fund_participation_constraints: Some(
