+++ rs/nns/integration_tests/src/reset_root.rs
@@ -5,7 +5,7 @@ use ic_nns_common::pb::v1::NeuronId;
 use ic_nns_constants::{LIFELINE_CANISTER_ID, ROOT_CANISTER_ID};
 use ic_nns_governance_api::{
     pb::v1::{manage_neuron_response::Command, NnsFunction},
-    proposal_helpers::create_external_update_proposal_candid,
+    proposal_submission_helpers::create_external_update_proposal_candid,
 };
 use ic_nns_test_utils::{
     common::{build_root_wasm, modify_wasm_bytes, NnsInitPayloadsBuilder},
