+++ rs/sns/cli/Cargo.toml
@@ -31,7 +31,7 @@ ic-nervous-system-proto = { path = "../../nervous_system/proto" }
 ic-nns-common = { path = "../../nns/common" }
 ic-nns-constants = { path = "../../nns/constants" }
 ic-nns-governance = { path = "../../nns/governance" }
-ic-sns-governance = { path = "../governance" }
+ic-nns-governance-api = { path = "../../nns/governance/api" }
 ic-sns-init = { path = "../init" }
 ic-sns-root = { path = "../root" }
 ic-sns-wasm = { path = "../../nns/sns-wasm" }
