+++ rs/sns/init/BUILD.bazel
@@ -33,12 +33,14 @@ BASE_DEPENDENCIES = [
 # Each target declared in this file may choose either these (release-ready)
 # dependencies (`DEPENDENCIES`), or `DEPENDENCIES_WITH_TEST_FEATURES` feature previews.
 DEPENDENCIES = BASE_DEPENDENCIES + [
+    "//rs/nns/governance/api",
     "//rs/nns/constants",
     "//rs/sns/governance",
     "//rs/sns/swap",
 ]
 
 DEPENDENCIES_WITH_TEST_FEATURES = BASE_DEPENDENCIES + [
+    "//rs/nns/governance/api:api--test_feature",
     "//rs/nns/constants:constants--test_feature",
     "//rs/sns/governance:governance--test_feature",
     "//rs/sns/swap:swap--test_feature",
